Add game to cart

+++ src/main/java/cringe/app/db/GameRepository.java
@@ -2,6 +2,6 @@ package cringe.app.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRepository extends JpaRepository<Game, Long> {
+public interface GameRepository extends JpaRepository<Game, Integer> {
     Game findGameById(int id);
 }
